proxy: support bookkeeping for unproxied object
